+++ test/Bitbucket/Tests/API/UserTest.php
@@ -7,6 +7,24 @@ use Bitbucket\API;
 
 class UserTest extends Tests\TestCase
 {
+    public function testGetEmails()
+    {
+        $endpoint       = 'user/emails';
+        $expectedResult = $this->fakeResponse(array('dummy'));
+
+        $client = $this->getHttpClientMock();
+        $client->expects($this->any())
+            ->method('get')
+            ->with($endpoint)
+            ->will($this->returnValue($expectedResult));
+
+        /** @var \Bitbucket\API\User $user */
+        $user = $this->getClassMock('Bitbucket\API\User', $client);
+        $actual = $user->emails();
+
+        $this->assertEquals($expectedResult, $actual);
+    }
+
     public function testGetUserProfileSuccess()
     {
         $endpoint       = 'user/';